32/3/requirements.txt
flask
pymongo
pyyaml